// heart-api-frontend/src/pages/account/center/data.d.ts
export type tabKeyType = 'articles' | 'applications' | 'projects';
export interface TagType {
  key: string;
  label: string;
}

export type GeographicType = {
  province: {
    label: string;
    key: string;
  };
  city: {
    label: string;
    key: string;
  };
};

export type NoticeType = {
  id: string;
  title: string;
  logo: string;
  description: string;
  updatedAt: string;
  member: string;
  href: string;
  memberLink: string;
};

export type CurrentUser = {
  // name: string;
  // avatar: string;
  // userid: string;
  // notice: NoticeType[];
  // email: string;
  // signature: string;
  // title: string;
  // group: string;
  // tags: TagType[];
  // notifyCount: number;
  // unreadCount: number;
  // country: string;
  // geographic: GeographicType;
  // address: string;
  // phone: string;
  id: number;
  userName: string;
  userAccount: string;
  userAvatar?: string;
  accessKey?: string;
  secretKey?: string
  signature?: string
  gender?: string;
  userRole: string;
  createTime: Date;
};

export type Member = {
  avatar: string;
  name: string;
  id: string;
};

export type ListItemDataType = {
  id: string;
  owner: string;
  title: string;
  avatar: string;
  cover: string;
  status: 'normal' | 'exception' | 'active' | 'success';
  percent: number;
  logo: string;
  href: string;
  body?: any;
  updatedAt: number;
  createdAt: number;
  subDescription: string;
  description: string;
  activeUser: number;
  newUser: number;
  star: number;
  like: number;
  message: number;
  content: string;
  members: Member[];
};